+++ src/config.h
@@ -251,6 +251,8 @@ typedef struct {
   wstring options_font;
   int options_fontsize;
   string old_options;
+  bool dim_margins;
+  bool status_line;
   bool old_xbuttons;
   // Legacy
   bool use_system_colours;