+++ src/term.c
@@ -332,6 +332,7 @@ void
 
   if (full) {
     term.lrmargmode = false;
+    term.dim_margins = cfg.dim_margins;
     term.deccolm_allowed = cfg.enable_deccolm_init;  // not reset by xterm
     term.vt220_keys = vt220(cfg.term);  // not reset by xterm
     term.app_keypad = false;  // xterm only with RIS
@@ -406,6 +407,11 @@ void
   term_switch_status(false);
   if (full) {
     term_clear_status();
+    // restore initial status line unless type 2 selected
+    if (cfg.status_line && term.st_type == 0)
+      term_set_status_type(1, 0);
+    else if (!cfg.status_line && term.st_type == 1)
+      term_set_status_type(0, 0);
   }
 
   if (full) {
@@ -1965,6 +1971,86 @@ void
 }
 
 /*
+ * Set status type.
+ */
+void
+(term_set_status_type)(struct term* term_p, int type, int lines)
+{
+  TERM_VAR_REF(true)
+  
+  /*
+    Unlike xterm, we do not resize the window when changing 
+    status area size; this does not only avoid trouble in full-screen 
+    or maximized mode, it also complies with DEC (VT420 p. 221, VT520 p. 5-147):
+	Notes on DECSSDT
+	• If you select no status line (Ps = 0), the terminal uses the 
+	  line as an additional user window line to display data.
+    As an extension, mintty supports a multi-line status area, 
+    configured with a second parameter to DECSSDT 2.
+    The suggestion of such an option could be interpreted from VT520 p. 2-35:
+	[The number of data display lines visible, not counting any status lines.]
+    although that may just be referring to status lines of multiple sessions.
+  */
+  int old_st_rows = term.st_rows;
+  switch (type) {
+    when 0: 
+            term_clear_status();
+            term.st_rows = 0;
+    when 1: 
+            if (term.st_type == 2)
+              term_clear_status();
+            term.st_type = 1; term.st_rows = 1;
+    when 2: 
+            if (term.st_type != 2)
+              term_clear_status();
+            term.st_type = 2;
+            if (lines) {
+              if (lines >= term_allrows / 2)
+                term.st_rows = max(0, term_allrows / 2 - 1);
+              else
+                term.st_rows = lines;
+            }
+            else
+              term.st_rows = 1;
+    othwise: return;
+  }
+  if (!term.st_rows) {
+    term.st_type = 0;
+    term_switch_status(false);
+  }
+  if (term.st_type != 2)
+    term_switch_status(false);
+  if (term.st_rows != old_st_rows) {
+    // don't need to win_adapt_term_size(false, false);
+    int newrows = term.rows + old_st_rows - term.st_rows;
+    // scroll cursor position out of status line area
+    int n = term.curs.y - newrows + 1;
+    if (n > 0) {
+      bool old_st_act = term.st_active;
+      term_switch_status(false);
+      term_do_scroll(term.marg_top, term.marg_bot, n, true);
+      // fix up
+      term.curs.y = max(0, term.curs.y - n);
+      term_switch_status(old_st_act);
+    }
+    // don't need to term_resize(newrows, -term.cols);
+    term.rows = newrows;
+    term.marg_top = 0;
+    term.marg_bot = newrows - 1;
+    term.marg_left = 0;
+    term.marg_right = term.cols - 1;
+    // clear status lines
+    for (int i = term.rows; i < term_allrows; i++) {
+      freeline(term.lines[i]);
+      term.lines[i] = newline(term.cols, false);
+    }
+    // notify child process
+    struct winsize ws = {(short unsigned)newrows, (short unsigned)term.cols, (short unsigned)(term.cols * cell_width), (short unsigned)(newrows * cell_height)};
+    child_resize(&ws);
+  }
+}
+
+/*
  * Swap active status line / main display.
  */
 void
@@ -1980,8 +2066,11 @@ void
   term_cursor oldcurs = term.curs;
   term.curs = term.st_other_curs;
   // saved status cursor line is normalized to 0, adjust
-  if (status_line)
+  if (status_line) {
     term.curs.y += term.rows;
+    if (term.curs.y >= term_allrows)
+      term.curs.y = term_allrows - 1;
+  }
   else
     oldcurs.y -= term.rows;
   term.st_other_curs = oldcurs;
@@ -3492,6 +3581,24 @@ void
       newchars[j].chr = tchar;
      /* Combining characters are still read from chars */
       newchars[j].cc_next = 0;
+
+     /* Margin indication */
+      if (term.dim_margins &&
+          (// in normal display:
+           i < term.marg_top || (i > term.marg_bot && i < term.rows)
+           // in status display:
+           || (i >= term.rows && i < term.marg_top + term.rows)
+           || (i > term.marg_bot + term.rows)
+           // outside left/right margins:
+           || j < term.marg_left || j > term.marg_right
+          )
+         )
+      {
+        cattr * ca = &newchars[j].attr;
+        ca->attr |= ATTR_DIM;
+        ca->attr ^= ATTR_REVERSE;
+      }
+
     }  // end first loop
 
     if (i == curs_y) {
@@ -3564,7 +3671,10 @@ void
           dispchars[curs_x].attr.attr |= ATTR_INVALID;
           curs_x++;
         }
+        term.st_kb_flag = what;
       }
+      else
+        term.st_kb_flag = 0;
 
      /* Progress indication */
       if (term.detect_progress) {