+++ src/term.h
@@ -575,6 +575,7 @@ struct term {
   int marg_top, marg_bot; /* scrolling region margins */
   int marg_left, marg_right; /* horizontal margins */
   bool lrmargmode;           /* enable horizontal margins */
+  bool dim_margins;
   bool attr_rect;            /* rectangular attribute change extent */
   bool printing, only_printing;  /* Are we doing ANSI printing? */
   int  print_state;       /* state of print-end-sequence scan */
@@ -591,6 +592,7 @@ struct term {
   bool st_active;         /* status line display active (DECSASD 1)*/
   term_cursor st_other_curs;  /* switched with active status display */
   term_cursor st_saved_curs;
+  uint st_kb_flag;
 
   bool has_focus;
   bool focus_reported;