+++ src/termout.c
@@ -297,6 +297,10 @@ static void
   TERM_VAR_REF(true)
   
   //printf("attr_rect %d,%d..%d,%d +%llX -%llX ^%llX\n", y0, x0, y1, x1, add, sub, xor);
+  if (term.st_active) {
+    y0 += term.rows;
+    y1 += term.rows;
+  }
   y0--; x0--; y1--; x1--;
 
   if (term.curs.origin) {
@@ -369,6 +373,10 @@ static void
   TERM_VAR_REF(true)
   
   //printf("fill_rect %d,%d..%d,%d\n", y0, x0, y1, x1);
+  if (term.st_active) {
+    y0 += term.rows;
+    y1 += term.rows;
+  }
   int width = charwidth(chr);
   if (chr == UCSWIDE || width < 1)
     return;
@@ -500,6 +508,11 @@ static void
   TERM_VAR_REF(true)
   
   //printf("copy_rect %d,%d..%d,%d -> %d,%d\n", y0, x0, y1, x1, y2, x2);
+  if (term.st_active) {
+    y0 += term.rows;
+    y1 += term.rows;
+    y2 += term.rows;
+  }
   y0--; x0--; y1--; x1--; y2--; x2--;
 
   if (term.curs.origin) {
@@ -650,6 +663,10 @@ static uint
   TERM_VAR_REF(true)
   
   //printf("sum_rect %d,%d..%d,%d\n", y0, x0, y1, x1);
+  if (term.st_active) {
+    y0 += term.rows;
+    y1 += term.rows;
+  }
 
   y0--; x0--; y1--; x1--;
 
@@ -2928,11 +2945,11 @@ static void
     }
     when 'd':        /* VPA: set vertical position */
       move(curs->x,
-           (curs->origin ? term.marg_top : 0) + arg0_def1 - 1,
+           (curs->origin ? term.marg_top : 0) + arg0_def1 + top_y - 1,
            curs->origin ? 2 : 0);
     when 'H' case_or 'f':  /* CUP or HVP: set horiz. and vert. positions at once */
       move((curs->origin ? term.marg_left : 0) + (arg1 ?: 1) - 1,
-           (curs->origin ? term.marg_top : 0) + arg0_def1 - 1,
+           (curs->origin ? term.marg_top : 0) + arg0_def1 + top_y - 1,
            curs->origin ? 2 : 0);
     when 'I':  /* CHT: move right N TABs */
       for (int i = 0; i < arg0_def1; i++)
@@ -3266,7 +3283,7 @@ static void
     when 'n':        /* DSR: device status report */
       if (arg0 == 6)  // CPR
         child_printf("\e[%d;%dR",
-                     curs->y + 1 - (curs->origin ? term.marg_top : 0),
+                     curs->y + 1 - (curs->origin ? term.marg_top : 0) - top_y,
                      curs->x + 1 - (curs->origin ? term.marg_left : 0));
       else if (arg0 == 5)
         child_write("\e[0n", 4);  // "in good operating condition"
@@ -3274,7 +3291,7 @@ static void
       switch (arg0) {
         when 6:  // DECXCPR
           child_printf("\e[?%d;%dR",  // VT420: third parameter "page"...
-                       curs->y + 1 - (curs->origin ? term.marg_top : 0),
+                       curs->y + 1 - (curs->origin ? term.marg_top : 0) - top_y,
                        curs->x + 1 - (curs->origin ? term.marg_left : 0));
         when 15:
           child_printf("\e[?%un", 11 - !!*cfg.printer);
@@ -3653,76 +3670,7 @@ static float freq_C5_C7[26] =
         }
     }
     when CPAIR('$', '~'): {  /* DECSSDT: select status line type */
-      /*
-        Unlike xterm, we do not resize the window when changing 
-        status area size; this does not only avoid trouble in full-screen 
-        or maximized mode, it also complies with DEC (VT420 p. 221, VT520 p. 5-147):
-	Notes on DECSSDT
-	• If you select no status line (Ps = 0), the terminal uses the 
-	  line as an additional user window line to display data.
-        As an extension, mintty supports a multi-line status area, 
-        configured with a second parameter to DECSSDT 2.
-        The suggestion of such an option could be interpreted from VT520 p. 2-35:
-	[The number of data display lines visible, not counting any status lines.]
-        although that may just be referring to status lines of multiple sessions.
-      */
-      int old_st_rows = term.st_rows;
-      switch (arg0) {
-        when 0: 
-                term_clear_status();
-                term.st_rows = 0;
-        when 1: 
-                if (term.st_type == 2)
-                  term_clear_status();
-                term.st_type = 1; term.st_rows = 1;
-        when 2: 
-                if (term.st_type != 2)
-                  term_clear_status();
-                term.st_type = 2;
-                if (arg1) {
-                  if (arg1 >= term_allrows / 2)
-                    term.st_rows = max(0, term_allrows / 2 - 1);
-                  else
-                    term.st_rows = arg1;
-                }
-                else
-                  term.st_rows = 1;
-        othwise: return;
-      }
-      if (!term.st_rows) {
-        term.st_type = 0;
-        term_switch_status(false);
-      }
-      if (term.st_type != 2)
-        term_switch_status(false);
-      if (term.st_rows != old_st_rows) {
-        // don't need to win_adapt_term_size(false, false);
-        int newrows = term.rows + old_st_rows - term.st_rows;
-        // scroll cursor position out of status line area
-        int n = curs->y - newrows + 1;
-        if (n > 0) {
-          bool old_st_act = term.st_active;
-          term_switch_status(false);
-          term_do_scroll(term.marg_top, term.marg_bot, n, true);
-          // fix up
-          curs->y = max(0, curs->y - n);
-          term_switch_status(old_st_act);
-        }
-        // don't need to term_resize(newrows, -term.cols);
-        term.rows = newrows;
-        term.marg_top = 0;
-        term.marg_bot = newrows - 1;
-        term.marg_left = 0;
-        term.marg_right = term.cols - 1;
-        // clear status lines
-        for (int i = term.rows; i < term_allrows; i++) {
-          freeline(term.lines[i]);
-          term.lines[i] = newline(term.cols, false);
-        }
-        // notify child process
-        struct winsize ws = {(short unsigned)newrows, (short unsigned)term.cols, (short unsigned)(term.cols * cell_width), (short unsigned)(newrows * cell_height)};
-        child_resize(&ws);
-      }
+      term_set_status_type(arg0, arg1);
     }
     when CPAIR('$', '}'): {  /* DECSASD: select active status display */
       bool status_line = arg0;
@@ -4370,7 +4318,7 @@ static void
     when 104: do_colour_osc(true, 4, true);
     when 105: do_colour_osc(true, 5, true);
     when 10:  do_colour_osc(false, FG_COLOUR_I, false);
-    when 11:  if (strchr("*_%=", *term.cmd_buf)) {
+    when 11:  if (strchr("*_%=+", *term.cmd_buf)) {
                 wchar * bn = cs__mbstowcs(term.cmd_buf);
                 wstrset(&cfg.background, bn);
                 free(bn);